+++ examples/millis/main.c
@@ -20,10 +20,10 @@ int main (void)
     printf("Testing millis()\n");
 
     for (;;)  {         
-        uint16_t prior_ticks = millis();
+        uint32_t prior_ticks = millis();
         delay(delay_time);
-        uint16_t delta_ticks = millis() - prior_ticks;
-        printf("Delay time was %u and delta in millis() was %u\n", delay_time, delta_ticks);
+        uint32_t delta_ticks = millis() - prior_ticks;
+        printf("Delay time was %u and delta in millis() was %lu\n", delay_time, delta_ticks);
     }
     return (0);
 }
